+++ apps/docs/app/docs/[...slug]/page.tsx
@@ -1,24 +1,14 @@
+import { Edit } from 'lucide-react';
 import type { Metadata } from 'next';
-import {
-  DocsPage,
-  DocsBody,
-  DocsTitle,
-  DocsDescription,
-  DocsCategory,
-} from 'fumadocs-ui/page';
+import { Card, Cards } from '@maximai/fumadocs-ui/components/card';
+import { DocsPage, DocsBody } from '@maximai/fumadocs-ui/page';
 import { notFound } from 'next/navigation';
-import { type ComponentProps, type FC, Fragment } from 'react';
-import defaultComponents from 'fumadocs-ui/mdx';
-import { Popup, PopupContent, PopupTrigger } from 'fumadocs-ui/twoslash/popup';
-import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
-import { Callout } from 'fumadocs-ui/components/callout';
-import { TypeTable } from 'fumadocs-ui/components/type-table';
-import { Accordion, Accordions } from 'fumadocs-ui/components/accordion';
-import { getImageMeta } from 'fumadocs-ui/og';
-import Preview from '@/components/preview';
+import { type ReactNode } from 'react';
+import { utils, type Page } from '@/utils/source';
 import { createMetadata } from '@/utils/metadata';
-import { openapi, utils } from '@/app/source';
-import { Wrapper } from '@/components/preview/wrapper';
+import Preview from '@/components/preview';
+import { cn } from '@/utils/cn';
+import { buttonVariants } from '@/components/ui/button';
 
 interface Param {
   slug: string[];
@@ -36,50 +26,79 @@ export default function Page({
   const path = `apps/docs/content/docs/${page.file.path}`;
   const preview = page.data.preview;
 
+  const footer = (
+    <a
+      href={`https://github.com/fuma-nama/fumadocs/blob/main/${path}`}
+      target="_blank"
+      rel="noreferrer noopener"
+      className={cn(
+        buttonVariants({
+          variant: 'secondary',
+          size: 'sm',
+          className: 'text-xs gap-1.5',
+        }),
+      )}
+    >
+      <Edit className="size-3" />
+      Edit on Github
+    </a>
+  );
+
   return (
     <DocsPage
-      toc={page.data.toc}
-      lastUpdate={page.data.lastModified}
+      toc={page.data.exports.toc}
+      lastUpdate={page.data.exports.lastModified}
       full={page.data.full}
       tableOfContent={{
-        style: 'clerk',
-        single: false,
+        footer,
       }}
-      editOnGithub={{
-        repo: 'fumadocs',
-        owner: 'fuma-nama',
-        sha: 'main',
-        path,
-      }}
+      tableOfContentPopover={{ footer }}
     >
-      <DocsTitle>{page.data.title}</DocsTitle>
-      <DocsDescription>{page.data.description}</DocsDescription>
+      <h1 className="text-3xl font-bold text-foreground sm:text-4xl">
+        {page.data.title}
+      </h1>
+      <p className="mb-8 text-lg text-muted-foreground">
+        {page.data.description}
+      </p>
       <DocsBody>
         {preview && preview in Preview ? Preview[preview] : null}
-        <page.data.body
+        <page.data.exports.default
           components={{
-            ...defaultComponents,
-            Popup,
-            PopupContent,
-            PopupTrigger,
-            Tabs,
-            Tab,
-            TypeTable,
-            Accordion,
-            Accordions,
-            Wrapper,
-            blockquote: Callout as unknown as FC<ComponentProps<'blockquote'>>,
-            APIPage: openapi.APIPage,
             HeadlessOnly:
-              params.slug[0] === 'headless' ? Fragment : () => undefined,
-            UIOnly: params.slug[0] === 'ui' ? Fragment : () => undefined,
+              params.slug[0] === 'headless'
+                ? ({ children }: { children: ReactNode }) => children
+                : () => undefined,
+            UIOnly:
+              params.slug[0] === 'ui'
+                ? ({ children }: { children: ReactNode }) => children
+                : () => undefined,
           }}
         />
-        {page.data.index ? (
-          <DocsCategory page={page} pages={utils.getPages()} />
-        ) : null}
+        {page.data.index ? <Category page={page} /> : null}
       </DocsBody>
     </DocsPage>
+  );
+}
+
+function Category({ page }: { page: Page }): React.ReactElement {
+  const filtered = utils
+    .getPages()
+    .filter(
+      (item) =>
+        item.file.dirname === page.file.dirname && item.file.name !== 'index',
+    );
+
+  return (
+    <Cards>
+      {filtered.map((item) => (
+        <Card
+          key={item.url}
+          title={item.data.title}
+          description={item.data.description ?? 'No Description'}
+          href={item.url}
+        />
+      ))}
+    </Cards>
   );
 }
 
@@ -91,7 +110,12 @@ export function generateMetadata({ params }: { params: Param }): Metadata {
   const description =
     page.data.description ?? 'The library for building documentation sites';
 
-  const image = getImageMeta('og', page.slugs);
+  const image = {
+    alt: 'Banner',
+    url: `/og/docs/${page.slugs.join('/')}.png`,
+    width: 1200,
+    height: 630,
+  };
 
   return createMetadata({
     title: page.data.title,
@@ -107,5 +131,7 @@ export function generateMetadata({ params }: { params: Param }): Metadata {
 }
 
 export function generateStaticParams(): Param[] {
-  return utils.generateParams();
+  return utils.getPages().map<Param>((page) => ({
+    slug: page.slugs,
+  }));
 }