+++ apps/docs/components/code-block.tsx
@@ -1,39 +1,54 @@
-import * as Base from 'fumadocs-ui/components/codeblock';
-import { Fragment } from 'react';
-import { codeToHast } from 'shiki';
-import { toJsxRuntime, type Jsx } from 'hast-util-to-jsx-runtime';
-import { jsx, jsxs } from 'react/jsx-runtime';
+import * as Base from '@maximai/fumadocs-ui/components/codeblock';
+import type { HTMLAttributes } from 'react';
+import { useMemo } from 'react';
+import { createHighlighter } from 'shiki';
 
-export interface CodeBlockProps {
+const highlighter = await createHighlighter({
+  langs: ['bash', 'ts', 'tsx'],
+  themes: ['github-light', 'github-dark'],
+});
+
+export type CodeBlockProps = HTMLAttributes<HTMLPreElement> & {
   code: string;
   wrapper?: Base.CodeBlockProps;
   lang: 'bash' | 'ts' | 'tsx';
-}
+};
 
-export async function CodeBlock({
+export function CodeBlock({
   code,
   lang,
   wrapper,
-}: CodeBlockProps): Promise<React.ReactElement> {
-  const hast = await codeToHast(code, {
-    lang,
-    defaultColor: false,
-    themes: {
-      light: 'github-light',
-      dark: 'github-dark',
-    },
-  });
+  ...props
+}: CodeBlockProps): React.ReactElement {
+  const html = useMemo(
+    () =>
+      highlighter.codeToHtml(code, {
+        lang,
+        defaultColor: false,
+        themes: {
+          light: 'github-light',
+          dark: 'github-dark',
+        },
+        transformers: [
+          {
+            name: 'remove-pre',
+            root: (root) => {
+              if (root.children[0].type !== 'element') return;
 
-  const rendered = toJsxRuntime(hast, {
-    jsx: jsx as Jsx,
-    jsxs: jsxs as Jsx,
-    Fragment,
-    development: false,
-    components: {
-      // @ts-expect-error -- JSX component
-      pre: Base.Pre,
-    },
-  });
+              return {
+                type: 'root',
+                children: root.children[0].children,
+              };
+            },
+          },
+        ],
+      }),
+    [code, lang],
+  );
 
-  return <Base.CodeBlock {...wrapper}>{rendered}</Base.CodeBlock>;
+  return (
+    <Base.CodeBlock {...wrapper}>
+      <Base.Pre {...props} dangerouslySetInnerHTML={{ __html: html }} />
+    </Base.CodeBlock>
+  );
 }