+++ apps/docs/components/preview/index.tsx
@@ -1,37 +1,57 @@
 import { HomeIcon } from 'lucide-react';
-import { Accordion, Accordions } from 'fumadocs-ui/components/accordion';
-import { Callout } from 'fumadocs-ui/components/callout';
-import { File, Folder, Files } from 'fumadocs-ui/components/files';
-import { Step, Steps } from 'fumadocs-ui/components/steps';
-import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
-import { TypeTable } from 'fumadocs-ui/components/type-table';
-import { Card } from 'fumadocs-ui/components/card';
-import { Heading } from 'fumadocs-ui/components/heading';
+import {
+  Accordion,
+  Accordions,
+} from '@maximai/fumadocs-ui/components/accordion';
+import { Callout } from '@maximai/fumadocs-ui/components/callout';
+import { File, Folder, Files } from '@maximai/fumadocs-ui/components/files';
+import { Step, Steps } from '@maximai/fumadocs-ui/components/steps';
+import { Tab, Tabs } from '@maximai/fumadocs-ui/components/tabs';
+import { TypeTable } from '@maximai/fumadocs-ui/components/type-table';
+import { Card } from '@maximai/fumadocs-ui/components/card';
+import { Heading } from '@maximai/fumadocs-ui/components/heading';
 import type { ReactNode } from 'react';
-import { RootToggle } from 'fumadocs-ui/components/layout/root-toggle';
-import { Banner } from 'fumadocs-ui/components/banner';
-import { InlineTOC } from 'fumadocs-ui/components/inline-toc';
-import { ImageZoom } from 'fumadocs-ui/components/image-zoom';
+import { RootToggle } from '@maximai/fumadocs-ui/components/layout/root-toggle';
+import dynamic from 'next/dynamic';
 import BannerImage from '@/public/banner.png';
 import { modes } from '@/utils/modes';
 import { Wrapper } from './wrapper';
 
+const RollButton = dynamic(() =>
+  import('@maximai/fumadocs-ui/components/roll-button').then(
+    (m) => m.RollButton,
+  ),
+);
+
+const Banner = dynamic(() =>
+  import('@maximai/fumadocs-ui/components/banner').then((m) => m.Banner),
+);
+
+const InlineTOC = dynamic(() =>
+  import('@maximai/fumadocs-ui/components/inline-toc').then(
+    (res) => res.InlineTOC,
+  ),
+);
+
+const ImageZoom = dynamic(() =>
+  import('@maximai/fumadocs-ui/components/image-zoom').then((m) => m.ImageZoom),
+);
+
 export default {
   heading: (
     <Wrapper>
-      <div className="rounded-lg bg-fd-background p-4 prose-no-margin">
-        <Heading id="preview" as="h3">
-          Hello World
-        </Heading>
-        <Heading id="preview" as="h3">
-          Hello <code>World</code> Everyone!
-        </Heading>
-      </div>
+      <Heading
+        id="preview"
+        as="h3"
+        className="!my-0 rounded-xl bg-background p-4"
+      >
+        Hello World
+      </Heading>
     </Wrapper>
   ),
   card: (
     <Wrapper>
-      <div className="rounded-lg bg-fd-background">
+      <div className="rounded-lg bg-background">
         <Card
           href="#"
           icon={<HomeIcon />}
@@ -43,9 +63,9 @@ export default {
   ),
   tabs: (
     <Wrapper>
-      <div className="space-y-4 rounded-xl bg-fd-background p-4 text-sm">
+      <div className="space-y-4 rounded-xl bg-background p-4 text-sm">
         <Tabs
-          groupId="language"
+          id="language"
           persist
           items={['Javascript', 'Rust', 'Typescript']}
         >
@@ -54,7 +74,7 @@ export default {
           <Tab value="Typescript">Also works if items are not the same</Tab>
         </Tabs>
 
-        <Tabs groupId="language" persist items={['Javascript', 'Rust']}>
+        <Tabs id="language" persist items={['Javascript', 'Rust']}>
           <Tab value="Javascript">
             Value is shared! Try refresh and see if the value is persisted
           </Tab>
@@ -67,7 +87,7 @@ export default {
   ),
   'type-table': (
     <Wrapper>
-      <div className="rounded-xl bg-fd-background px-4">
+      <div className="rounded-xl bg-background px-4">
         <TypeTable
           type={{
             percentage: {
@@ -86,7 +106,7 @@ export default {
       <ImageZoom
         alt="banner"
         src={BannerImage}
-        className="!my-0 rounded-xl bg-fd-background"
+        className="!my-0 rounded-xl bg-background"
         priority
       />
     </Wrapper>
@@ -123,7 +143,6 @@ export default {
           <File name="button.tsx" />
           <File name="tabs.tsx" />
           <File name="dialog.tsx" />
-          <Folder name="empty" />
         </Folder>
         <File name="package.json" />
       </Files>
@@ -179,7 +198,7 @@ export default {
   ),
   steps: (
     <Wrapper>
-      <div className="rounded-xl bg-fd-background p-3">
+      <div className="rounded-xl bg-background p-3">
         <Steps>
           <Step>
             <h4>Buy Coffee</h4>
@@ -197,9 +216,15 @@ export default {
       </div>
     </Wrapper>
   ),
+  'roll-button': (
+    <div className="rounded-lg border bg-card p-4 text-muted-foreground">
+      <RollButton />
+      <p className="text-center text-sm">Scroll down to see the button</p>
+    </div>
+  ),
   'root-toggle': (
     <Wrapper>
-      <div className="not-prose rounded-xl bg-fd-background p-3">
+      <div className="not-prose rounded-xl bg-background p-3">
         <RootToggle
           options={modes.map((mode) => ({
             url: `/docs/${mode.param}`,
@@ -214,22 +239,8 @@ export default {
   banner: (
     <Wrapper>
       <div className="flex flex-col gap-4">
-        <Banner className="z-0" changeLayout={false}>
-          Be careful, Fumadocs v99 has released
-        </Banner>
-        <Banner
-          className="z-0"
-          id="test-rainbow"
-          variant="rainbow"
-          changeLayout={false}
-        >
-          <span className="opacity-90 mix-blend-luminosity">
-            Using the <code>rainbow</code> variant
-          </span>
-        </Banner>
-        <Banner className="z-0" id="test" changeLayout={false}>
-          Be careful, this banner can be closed
-        </Banner>
+        <Banner>Be careful, Fumadocs v99 has released</Banner>
+        <Banner id="99">Be careful, this banner can be closed</Banner>
       </div>
     </Wrapper>
   ),