+++ apps/docs/content/docs/example/tailwind.config.js
@@ -1,11 +1,11 @@
-import { createPreset } from 'fumadocs-ui/tailwind-plugin';
+import { createPreset } from '@maximai/fumadocs-ui/tailwind-plugin';
 
 /** @type {import('tailwindcss').Config} */
 export default {
   darkMode: 'class',
   presets: [createPreset()],
   content: [
-    './node_modules/fumadocs-ui/dist/**/*.js',
+    './node_modules/@maximai/fumadocs-ui/dist/**/*.js',
 
     './components/**/*.{ts,tsx}',
     './app/**/*.{ts,tsx}',