+++ apps/docs/content/docs/ui/props.ts
@@ -1,13 +1,24 @@
-import type { Accordion, Accordions } from 'fumadocs-ui/components/accordion';
-import type { Callout } from 'fumadocs-ui/components/callout';
-import type { File, Folder } from 'fumadocs-ui/components/files';
-import type { InlineTOC } from 'fumadocs-ui/components/inline-toc';
-import type { TypeTable } from 'fumadocs-ui/components/type-table';
-import type { Card } from 'fumadocs-ui/components/card';
-import type { DocsLayoutProps } from 'fumadocs-ui/layout';
+import type {
+  Accordion,
+  Accordions,
+} from '@maximai/fumadocs-ui/components/accordion';
+import type { Callout } from '@maximai/fumadocs-ui/components/callout';
+import type {
+  API,
+  APIExample,
+  APIInfo,
+  Property,
+  Root,
+} from '@maximai/fumadocs-ui/components/api';
+import type { File, Folder } from '@maximai/fumadocs-ui/components/files';
+import type { InlineTOC } from '@maximai/fumadocs-ui/components/inline-toc';
+import type { RollButton } from '@maximai/fumadocs-ui/components/roll-button';
+import type { TypeTable } from '@maximai/fumadocs-ui/components/type-table';
+import type { Card } from '@maximai/fumadocs-ui/components/card';
+import type { DocsLayoutProps } from '@maximai/fumadocs-ui/layout';
 import type { AnchorHTMLAttributes, ComponentPropsWithoutRef } from 'react';
-import type { DocsPageProps } from 'fumadocs-ui/page';
-import type { AutoTypeTable } from 'fumadocs-typescript/ui';
+import type { DocsPageProps } from '@maximai/fumadocs-ui/page';
+import type { AutoTypeTable } from '@maximai/fumadocs-typescript/ui';
 
 export type AccordionsProps = Omit<
   ComponentPropsWithoutRef<typeof Accordions>,
@@ -16,6 +27,31 @@ export type AccordionsProps = Omit<
 
 export type AccordionProps = Omit<
   ComponentPropsWithoutRef<typeof Accordion>,
+  keyof ComponentPropsWithoutRef<'div'>
+>;
+
+export type RootProps = Omit<
+  ComponentPropsWithoutRef<typeof Root>,
+  keyof ComponentPropsWithoutRef<'div'>
+>;
+
+export type ApiProps = Omit<
+  ComponentPropsWithoutRef<typeof API>,
+  keyof ComponentPropsWithoutRef<'div'>
+>;
+
+export type ApiExampleProps = Omit<
+  ComponentPropsWithoutRef<typeof APIExample>,
+  keyof ComponentPropsWithoutRef<'div'>
+>;
+
+export type APIInfoProps = Omit<
+  ComponentPropsWithoutRef<typeof APIInfo>,
+  keyof ComponentPropsWithoutRef<'div'>
+>;
+
+export type PropertyProps = Omit<
+  ComponentPropsWithoutRef<typeof Property>,
   keyof ComponentPropsWithoutRef<'div'>
 >;
 
@@ -43,6 +79,8 @@ export type CardProps = Omit<
   ComponentPropsWithoutRef<typeof Card>,
   keyof Omit<AnchorHTMLAttributes<HTMLAnchorElement>, 'href'>
 >;
+
+export type RollButtonProps = ComponentPropsWithoutRef<typeof RollButton>;
 
 export type TypeTableProps = ComponentPropsWithoutRef<typeof TypeTable>;