+++ examples/content-collections/app/docs/[[...slug]]/page.tsx
@@ -1,21 +1,16 @@
-import { source } from '@/app/source';
+import { getPage, getPages } from '@/app/source';
 import type { Metadata } from 'next';
-import {
-  DocsPage,
-  DocsBody,
-  DocsTitle,
-  DocsDescription,
-} from 'fumadocs-ui/page';
+import { DocsPage, DocsBody } from '@maximai/fumadocs-ui/page';
 import { notFound } from 'next/navigation';
 import { MDXContent } from '@content-collections/mdx/react';
-import defaultMdxComponents from 'fumadocs-ui/mdx';
+import defaultMdxComponents from '@maximai/fumadocs-ui/mdx';
 
 export default async function Page({
   params,
 }: {
   params: { slug?: string[] };
 }) {
-  const page = source.getPage(params.slug);
+  const page = getPage(params.slug);
 
   if (!page) {
     notFound();
@@ -23,9 +18,8 @@ export default async function Page({
 
   return (
     <DocsPage toc={page.data.toc} full={page.data.full}>
-      <DocsTitle>{page.data.title}</DocsTitle>
-      <DocsDescription>{page.data.description}</DocsDescription>
       <DocsBody>
+        <h1>{page.data.title}</h1>
         <MDXContent
           code={page.data.body}
           components={{ ...defaultMdxComponents }}
@@ -36,11 +30,13 @@ export default async function Page({
 }
 
 export async function generateStaticParams() {
-  return source.generateParams();
+  return getPages().map((page) => ({
+    slug: page.slugs,
+  }));
 }
 
 export function generateMetadata({ params }: { params: { slug?: string[] } }) {
-  const page = source.getPage(params.slug);
+  const page = getPage(params.slug);
 
   if (!page) notFound();