+++ packages/create-app/template/+shared/app/docs/layout.tsx
@@ -1,4 +1,4 @@
-import { DocsLayout } from 'fumadocs-ui/layout';
+import { DocsLayout } from '@maximai/fumadocs-ui/layout';
 import type { ReactNode } from 'react';
 import { baseOptions } from '../layout.config';
 import { source } from '@/app/source';