+++ packages/create-app/template/content-collections/app/docs/[[...slug]]/page.tsx
@@ -1,24 +1,19 @@
-import { source } from '@/app/source';
+import { getPage, getPages } from '@/app/source';
 import type { Metadata } from 'next';
-import {
-  DocsPage,
-  DocsBody,
-  DocsTitle,
-  DocsDescription,
-} from 'fumadocs-ui/page';
+import { DocsPage, DocsBody } from '@maximai/fumadocs-ui/page';
 import { notFound } from 'next/navigation';
 import { MDXContent } from '@content-collections/mdx/react';
-import defaultMdxComponents from 'fumadocs-ui/mdx';
+import defaultMdxComponents from '@maximai/fumadocs-ui/mdx';
 
 export default function Page({ params }: { params: { slug?: string[] } }) {
-  const page = source.getPage(params.slug);
+  const page = getPage(params.slug);
+
   if (!page) notFound();
 
   return (
     <DocsPage toc={page.data.toc} full={page.data.full}>
-      <DocsTitle>{page.data.title}</DocsTitle>
-      <DocsDescription>{page.data.description}</DocsDescription>
       <DocsBody>
+        <h1>{page.data.title}</h1>
         <MDXContent
           code={page.data.body}
           components={{ ...defaultMdxComponents }}
@@ -29,11 +24,14 @@ export default function Page({ params }: { params: { slug?: string[] } }) {
 }
 
 export function generateStaticParams() {
-  return source.generateParams();
+  return getPages().map((page) => ({
+    slug: page.slugs,
+  }));
 }
 
 export function generateMetadata({ params }: { params: { slug?: string[] } }) {
-  const page = source.getPage(params.slug);
+  const page = getPage(params.slug);
+
   if (!page) notFound();
 
   return {