+++ packages/doc-gen/src/typescript-generator.ts
@@ -7,32 +7,23 @@ import {
   generateDocumentation,
   renderMarkdownToHast,
   type DocEntry,
-} from 'fumadocs-typescript';
-import { getProgram } from 'fumadocs-typescript';
-import type { Program } from 'typescript';
-import { z } from 'zod';
+} from '@maximai/fumadocs-typescript';
 import { createElement, expressionToAttribute } from './utils';
 import type { DocGenerator } from './remark-docgen';
 
 export type TypescriptGeneratorOptions = GenerateDocumentationOptions;
 
-export type TypescriptGeneratorInput = z.output<
-  typeof typescriptGeneratorSchema
->;
-
-export const typescriptGeneratorSchema = z.object({
-  file: z.string({ description: 'Target TypeScript file name' }),
-  name: z.string({ description: 'Exported type name' }),
+export interface TypescriptGeneratorInput {
+  file: string;
+  name: string;
 
   /**
    * Component name which accepts the `type` property
    *
    * @defaultValue 'TypeTable'
    */
-  component: z
-    .string({ description: 'Component name which accepts the `type` property' })
-    .default('TypeTable'),
-});
+  component?: string;
+}
 
 export interface VirtualTypeTableProps {
   type: Record<
@@ -56,16 +47,8 @@ export interface VirtualTypeTableProps {
  * @param options - configuration
  */
 export function typescriptGenerator(
-  options: TypescriptGeneratorOptions = {},
+  options?: TypescriptGeneratorOptions,
 ): DocGenerator {
-  let program: Program | undefined;
-
-  function loadProgram(): Program {
-    return options.config && 'program' in options.config
-      ? options.config.program
-      : getProgram(options.config);
-  }
-
   function mapProperty(entry: DocEntry): Property {
     const value = valueToEstree({
       type: entry.type,
@@ -108,23 +91,15 @@ export function typescriptGenerator(
 
   return {
     name: 'typescript',
-    onFile() {
-      if (process.env.NODE_ENV === 'development' || !program) {
-        program = loadProgram();
-      }
-    },
     run(input, ctx) {
       const {
         file,
         name,
         component = 'TypeTable',
-      } = typescriptGeneratorSchema.parse(input);
+      } = input as TypescriptGeneratorInput;
       const dest = path.resolve(ctx.cwd, file);
 
-      const doc = generateDocumentation(dest, name, {
-        ...options,
-        config: { program: program ?? loadProgram() },
-      });
+      const doc = generateDocumentation(dest, name, options);
       if (!doc) throw new Error(`Failed to find type ${name} in ${dest}`);
 
       return createElement(component, [